cmov: have test macro write modules
Collapses the module structure by having the test macro declare the
module as well
